bugfix: update siibra-python nifti vol src chore: added test to catch siibra python change
